+++ CHANGELOG.md
@@ -3,6 +3,7 @@ vNext (Month Day, Year)
 
 **New this week**
 
+- (When complete) Add Event Stream support (#653, #xyz)
 - (When complete) Add profile file provider for region (#594, #xyz)
 
 v0.21 (August 19th, 2021)